CSS: Code display
I noticed that stackoverflow and some other sites use
a combination of pre-formatted text and code tags.

I like this idea. I've extended it to the notion of
multiline and (inline) non-multiline sections

Care must still be taken to convert HTML escape chars

I've no doubt this can be improved in time